适配问题修改 4

+++ beseLib/src/main/java/com/bese/widget/badge/BadgeView.kt
@@ -443,8 +443,8 @@ class BadgeView constructor(context: Context) : View(context) {
         return mBadgeTextColor
     }
 
-    fun setBadgeTextSize(size: Float): BadgeView {
-        mBadgeTextSize = size
+    fun setBadgeTextSize(size: Int): BadgeView {
+        mBadgeTextSize = size.toFloat()
         measureText()
         invalidate()
         return this
